line height to symbol graphics item

+++ src/common/graphictypes/musicfont/musicfont.cpp
@@ -115,7 +115,7 @@ void MusicFont::setLineHeight(int lineHeight)
     }
 }
 
-qreal MusicFont::lineHeight() const
+int MusicFont::lineHeight() const
 {
     return m_lineHeight;
 }
